+++ bindings/python/Python.cc
@@ -36,6 +36,7 @@ public:
                    bool spacer_annotate,
                    bool case_feature,
                    bool no_substitution,
+                   bool preserve_placeholders,
                    bool segment_case,
                    bool segment_numbers,
                    bool segment_alphabet_change,
@@ -52,6 +53,8 @@ public:
       flags |= onmt::Tokenizer::Flags::CaseFeature;
     if (no_substitution)
       flags |= onmt::Tokenizer::Flags::NoSubstitution;
+    if (preserve_placeholders)
+      flags |= onmt::Tokenizer::Flags::PreservePlaceholders;
     if (segment_case)
       flags |= onmt::Tokenizer::Flags::SegmentCase;
     if (segment_numbers)
@@ -126,7 +129,7 @@ BOOST_PYTHON_MODULE(pyonmttok)
 {
   py::class_<TokenizerWrapper>(
       "Tokenizer",
-      py::init<std::string, std::string, std::string, std::string, bool, bool, bool, bool, bool, bool, bool, bool, py::list>(
+      py::init<std::string, std::string, std::string, std::string, bool, bool, bool, bool, bool, bool, bool, bool, bool, py::list>(
         (py::arg("bpe_model_path")="",
          py::arg("sp_model_path")="",
          py::arg("joiner")=onmt::Tokenizer::joiner_marker,
@@ -135,6 +138,7 @@ BOOST_PYTHON_MODULE(pyonmttok)
          py::arg("spacer_annotate")=false,
          py::arg("case_feature")=false,
          py::arg("no_substitution")=false,
+         py::arg("preserve_placeholders")=false,
          py::arg("segment_case")=false,
          py::arg("segment_numbers")=false,
          py::arg("segment_alphabet_change")=false,